+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 1. 兼容 eslint v7.x 版本
 2. `@eslint-sets/eslint-config-all` 更改为 `@eslint-sets/eslint-config`
 3. `@nuxtjs/eslint-config` 升级到 v12
+4. 更改`all`包含 react、vue、ts、basic
 
 ## 2023.01.03 v3.3.0 更新日志